+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 ## 2026-02-19
+### Commit  (Introduce RabbitMQ-based messaging communication in the Core microservice)
+- Basic configuration and queue implementation test
 ### Commit 35a6ab1 (Associated with Gateway microservice)
 - Configure as Gateway client
 - Add `runMicroserviceInstance.sh` script to allow running multiple instances of this microservice simultaneously.